EntityNotFoundException was moved up to the service layer and implemented by standard null check instead of AOP.
