+++ child-service/src/main/java/com/plm/service/child/domain/ChildServiceImpl.java
@@ -1,6 +1,8 @@
 package com.plm.service.child.domain;
 
+import com.plm.service.child.dao.ChildEntity;
 import com.plm.service.child.dao.ChildEntityRepository;
+import com.plm.service.common.domain.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +18,12 @@ class ChildServiceImpl implements ChildService {
 
     @Override
     public Child getChildById(int id) {
-        return new Child(childEntityRepository.findOne(id));
+        ChildEntity childEntity = childEntityRepository.findOne(id);
+
+        if (childEntity != null) {
+            return new Child(childEntity);
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
 }
