+++ common-service/src/main/java/com/plm/service/common/domain/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package com.plm.service.common.dao;
+package com.plm.service.common.domain;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
